backend/requirements.txt
pandas
numpy
streamlit
groq

flask
flask-cors
requests
beautifulsoup4
newspaper3k
lxml
lxml_html_clean
pymongo
plotly
